refactor vue node browser tests by behaviors
